Day 2
Improved my organization for this one.

String parsing is always nice and easy for Python.

Going rework my template using the code here, I prefer this layout,
and it's clear now that the inputs will likely be the same for most
problems, meaning I can use the same parsing code