Day 10
This is the first day that really felt like I needed to think about data
structures. The previous days felt more like mind benders, instead of
specifically relying on knowledge of different structures.

Initially, I thought about using a tree, but realized that the
construction of this problem would make constructing a tree difficult,
however, that made me realize that an array would be perfect. I did it
with a Dict because it simplifies the logic and looks neater, but this
could be done with an array of length (input.max())